refactor(storage,runtime): W11-P1-D — AiSuccessOutcome token/latency 升 i64
按 storage-multi-dialect §4.2 ⚠️ 行，DTO 与读模型类型对齐。

读模型 ArticleAiResult.tokens_in/out/latency_ms 是 Option<i64>，
但写入 DTO AiSuccessOutcome 用 Option<i32>，三层（读模型 / DTO /
PG 列）类型不一致。统一升 DTO 为 Option<i64>：

- crates/storage/src/repo/article_ai_result.rs:31-35
  AiSuccessOutcome.tokens_in/out/latency_ms: Option<i32> → Option<i64>
- crates/runtime/src/flows/ai_run.rs
  clamp_u64_to_i32 → clamp_u64_to_i64（u64::MAX > i64::MAX 仍需饱和）
- 3 处调用点自动 follow rename

PG 列锁定 BIGINT（P2 schema 翻译时按读模型选）。SQLite 端 INTEGER
动态类型，bind 从 i32 升 i64 无行为差。

CI 三绿：fmt + clippy -D warnings + workspace 全测试套件通过（含
article_ai_advance_tests / article_ai_result_repo_tests 等 DTO
绑定路径）。

+++ crates/storage/src/repo/article_ai_result.rs
@@ -28,10 +28,10 @@ pub struct AiSuccessOutcome {
     pub importance_score: Option<i32>,
     pub keep_decision: Option<bool>,
     pub raw_response_artifact_id: Option<i64>,
-    pub tokens_in: Option<i32>,
-    pub tokens_out: Option<i32>,
+    pub tokens_in: Option<i64>,
+    pub tokens_out: Option<i64>,
     pub cost_micro_usd: Option<i64>,
-    pub latency_ms: Option<i32>,
+    pub latency_ms: Option<i64>,
 }
 
 #[derive(Debug, Clone)]
